requirements.txt
annotated-types==0.7.0
anyio==4.13.0
cffi==2.0.0
certifi==2026.2.25
charset-normalizer==3.4.7
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
filetype==1.2.0
google-auth==2.49.2
google-genai==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
langchain-core==1.2.31
langchain-google-genai==4.2.2
langchain-openai==1.1.14
langsmith==0.7.32
openai==2.32.0
orjson==3.11.8
packaging==26.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pydantic==2.13.1
pydantic_core==2.46.1
pycparser==3.0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
websockets==16.0
xxhash==3.6.0
zstandard==0.25.0
